refactor(opencode): store plugin body as a checked-in file embedded into the binary
