+++ station/src/bin/embeddedAssets.d.ts
@@ -7,3 +7,8 @@ declare module "*piExtension.mjs" {
   const embeddedPath: string;
   export default embeddedPath;
 }
+
+declare module "*station-opencode-plugin-body.js" {
+  const embeddedPath: string;
+  export default embeddedPath;
+}
